LPS-201156 Move LayoutPrototypePermissionUtil into Portal-impl
# breaking_change_report

## What portal-kernel/src/com/liferay/portal/kernel/service/permission/LayoutPrototypePermissionUtil.java

LayoutPrototypePermissionUtil is moving from portal-kernel into portal-impl.

## Why

We are merging portal-kernel into portal-impl as a long term goal to simplify portal structure.

## Alternatives

Add portal-impl in build dependency and use the same class.

----
