+++ modules/util/source-formatter/src/main/resources/dependencies/imports.txt
@@ -1029,7 +1029,7 @@ com.liferay.portal.service.permission.LayoutBranchPermissionUtil=com.liferay.por
 com.liferay.portal.service.permission.LayoutPermission=com.liferay.portal.kernel.service.permission.LayoutPermission
 com.liferay.portal.service.permission.LayoutPermissionUtil=com.liferay.portal.kernel.service.permission.LayoutPermissionUtil
 com.liferay.portal.service.permission.LayoutPrototypePermission=com.liferay.portal.kernel.service.permission.LayoutPrototypePermission
-com.liferay.portal.service.permission.LayoutPrototypePermissionUtil=com.liferay.portal.kernel.service.permission.LayoutPrototypePermissionUtil
+com.liferay.portal.service.permission.LayoutPrototypePermissionUtil=com.liferay.portal.service.permission.LayoutPrototypePermissionUtil
 com.liferay.portal.service.permission.LayoutSetBranchPermission=com.liferay.portal.kernel.service.permission.LayoutSetBranchPermission
 com.liferay.portal.service.permission.LayoutSetBranchPermissionUtil=com.liferay.portal.kernel.service.permission.LayoutSetBranchPermissionUtil
 com.liferay.portal.service.permission.LayoutSetPrototypePermission=com.liferay.portal.kernel.service.permission.LayoutSetPrototypePermission
